-Implemented signal graph loading across multithreading -Removed memory leak -turned datafile into separate class -removed bad QThread subclassing and changed to worker object running in QThread

+++ src/getTimeMs64.cpp
@@ -21,7 +21,7 @@ int GetTimeMs64()
      li.LowPart = ft.dwLowDateTime;
      li.HighPart = ft.dwHighDateTime;
 
-     uint ret = li.QuadPart;
+     long int ret = li.QuadPart;
      ret -= 116444736000000000LL; /* Convert from file time to UNIX epoch time. */
      ret /= 10000; /* From 100 nano seconds (10^-7) to 1 millisecond (10^-3) intervals */
